+++ includes/class-admin.php
@@ -313,6 +313,25 @@ class Admin {
 			)
 		);
 
+		\register_setting(
+			'activitypub',
+			'activitypub_mailer_new_follower',
+			array(
+				'type'        => 'boolean',
+				'description' => \__( 'Send notifications via e-mail when a new follower is added.', 'activitypub' ),
+				'default'     => '0',
+			)
+		);
+		\register_setting(
+			'activitypub',
+			'activitypub_mailer_new_dm',
+			array(
+				'type'        => 'boolean',
+				'description' => \__( 'Send notifications via e-mail when a direct message is received.', 'activitypub' ),
+				'default'     => '0',
+			)
+		);
+
 		// Blog-User Settings.
 		\register_setting(
 			'activitypub_blog',