+++ CHANGELOG.md
@@ -5,6 +5,44 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* Added a filter to make custom comment types manageable in WP.com Calypso
+
+### Fixed
+
+* Undefined array key warnings in various places
+
+## [4.6.0] - 2024-12-20
+
+### Added
+
+* Add a filter to allow modifying the ActivityPub preview template
+* `@mentions` in the JSON representation of the reply
+* Add settings to enable/disable e-mail notifications for new followers and direct messages
+
+### Changed
+
+* Direct Messages: Test for the user being in the to field
+* Direct Messages: Improve HTML to e-mail text conversion
+* Better support for FSE color schemes
+
+### Fixed
+
+* Reactions: Provide a fallback for empty avatar URLs
+
+## [4.5.1] - 2024-12-18
+
+### Changed
+
+* Reactions block: Remove the `wp-block-editor` dependency for frontend views
+
+### Fixed
+
+* Direct Messages: Don't send notification for received public activities
+
 ## [4.5.0] - 2024-12-17
 
 ### Added
@@ -16,7 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Account for custom comment types when calculating comment counts
 * Plugin upgrade routine that automatically updates comment counts
 
-### Improved
+### Changed
 
 * Likes and Reposts enabled by default
 * Email templates for Likes and Reposts
@@ -36,7 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Setting to enable/disable Authorized-Fetch
 
-### Improved
+### Changed
 
 * Added screen reader text to the "Follow Me" block for improved accessibility
 * Added `media_type` support to Activity-Object-Transformers
@@ -59,7 +97,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * GitHub action to enforce Changelog updates
 * New contributors
 
-### Improved
+### Changed
 
 * Basic enclosure validation
 * More User -> Actor renaming
@@ -82,7 +120,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Mastodon Apps status provider
 
-### Improved
+### Changed
 
 * Image-Handling
 * Have better checks if audience should be set or not
@@ -98,7 +136,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Unit tests for the `ActivityPub\Transformer\Post` class
 
-### Improved
+### Changed
 
 * Reuse constants once they're defined
 * "FEP-b2b8: Long-form Text" support
@@ -134,7 +172,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Hide sticky posts that are not public
 
-### Improved
+### Changed
 
 * `activity_handle_undo` action
 * Add title to content if post is a `Note`
@@ -146,7 +184,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Do not federate "Local" posts
 
-### Improved
+### Changed
 
 * Help-text for Content-Warning box
 
@@ -158,7 +196,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Seriously Simple Podcasting integration
 * Multiple small fixes
 
-### Improved
+### Changed
 
 * Provide contextual fallback for dynamic blocks
 
@@ -172,7 +210,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Post "visibility" feature
 * Attribution-Domains support
 
-### Improved
+### Changed
 
 * Inbox validation
 * WordPress-Post-Type - Detection
@@ -192,7 +230,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Sanitization callback
 
-### Improved
+### Changed
 
 * A lot of PHPCS cleanups
 * Prepare multi-lang support
@@ -204,7 +242,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Keep priority of Icons
 * Fatal error if remote-object is `WP_Error`
 
-### Improved
+### Changed
 
 * Adopt WordPress PHP Coding Standards
 
@@ -228,7 +266,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Cropping Header Images for users without the 'customize' capability
 
-### Improved
+### Changed
 
 * OpenSSL handling
 * Added missing @ in Follow-Me block
@@ -242,7 +280,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.2.4] - 2024-09-16
 
-### Improved
+### Changed
 
 * Inbox validation
 
@@ -253,7 +291,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * NodeInfo endpoint
 * (Temporarily) Remove HTML from `summary`, because it seems that Mastodon has issues with it
 
-### Improved
+### Changed
 
 * Accessibility for Reply-Context
 * Use `Article` Object-Type as default
@@ -282,7 +320,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * "Stream" Plugin support
 * New Fediverse symbol
 
-### Improved
+### Changed
 
 * Replace hashtags, URLs, and mentions in summary with links
 * Hide Bookmarklet if site does not support Blocks
@@ -303,7 +341,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Line breaks to user biography
 * Blueprint
 
-### Improved
+### Changed
 
 * Simplified WebFinger code
 
@@ -323,7 +361,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * ActivityPub link HTML/HTTP-Header support
 * Tag support for Actors (only auto-generated for now)
 
-### Improved
+### Changed
 
 * Add setting to enable/disable the `fediverse:creator` OGP tag.
 
@@ -344,7 +382,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Support for FEP-fb2a
 * CRUD support for Extra Fields
 
-### Improved
+### Changed
 
 * Remote-Follow UI and UX
 * Open Graph `fediverse:creator` implementation
@@ -525,7 +563,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Normalize attributes that can have mixed value types
 
-## [1.3.0] 2023-12-05
+## [1.3.0] - 2023-12-05
 
 ### Added
 
@@ -1150,8 +1188,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * initial
 
-[Unreleased]: https://github.com/Automattic/wordpress-activitypub/compare/4.5.0...trunk
+[Unreleased]: https://github.com/Automattic/wordpress-activitypub/compare/4.6.0...trunk
 <!-- Add new release below and update "Unreleased" link -->
+[4.6.0]: https://github.com/Automattic/wordpress-activitypub/compare/4.5.1...4.6.0
+[4.5.1]: https://github.com/Automattic/wordpress-activitypub/compare/4.5.0...4.5.1
 [4.5.0]: https://github.com/Automattic/wordpress-activitypub/compare/4.4.0...4.5.0
 [4.4.0]: https://github.com/Automattic/wordpress-activitypub/compare/4.3.0...4.4.0
 [4.3.0]: https://github.com/Automattic/wordpress-activitypub/compare/4.2.1...4.3.0